Down to the bare minimum, cannot please jest
