+++ package.json
@@ -4,8 +4,8 @@
   "description": "Consolidated parsing for languages with first-party support in Aerie sequencing",
   "type": "module",
   "scripts": {
-    "test": "npx jest",
-    "build": "tsc",
+    "test": "NODE_OPTIONS=--experimental-vm-modules npx jest",
+    "build": "rollup -c",
     "format": "prettier --write ./src"
   },
   "license": "MIT",
@@ -15,10 +15,16 @@
   },
   "devDependencies": {
     "@lezer/generator": "^1.7.3",
+    "@rollup/plugin-node-resolve": "^16.0.1",
+    "@rollup/plugin-typescript": "^12.1.2",
     "@types/jest": "^29.5.14",
+    "@types/node": "^22.14.1",
+    "jest": "^29.7.0",
     "prettier": "^3.5.3",
     "rollup": "^4.39.0",
-    "ts-jest": "^29.3.1",
+    "ts-jest": "^29.3.2",
+    "ts-node": "^10.9.2",
+    "tslib": "^2.8.1",
     "typescript": "^5.8.2"
   },
   "dependencies": {
@@ -26,8 +32,8 @@
     "@lezer/highlight": "^1.2.1",
     "@lezer/lr": "^1.4.2",
     "@nasa-jpl/aerie-ampcs": "^1.0.6",
+    "@nasa-jpl/aerie-time-utils": "git+https://github.com/NASA-AMMOS/aerie-time-utils.git",
     "@nasa-jpl/seq-json-schema": "^1.3.1",
-    "codemirror": "^6.0.1",
-    "@nasa-jpl/aerie-time-utils": "git+https://github.com/NASA-AMMOS/aerie-time-utils.git"
+    "codemirror": "^6.0.1"
   }
 }
