+++ rollup.config.js
@@ -1,14 +1,20 @@
-import {lezer} from "@lezer/generator/rollup"
+import typescript from '@rollup/plugin-typescript';
+import { lezer } from '@lezer/generator/rollup';
 
 export default {
-  input: ".build/src/index.js",
-  output: [{
-    format: "es",
-    file: "./dist/index.js"
-  }, {
-    format: "cjs",
-    file: "./dist/index.cjs"
-  }],
-  external: ["@lezer/lr"],
-  plugins: [lezer()]
-}
+  input: './src/index.ts',
+  output: [
+    {
+      format: 'es',
+      file: './dist/index.js',
+      sourcemap: true,
+    },
+    {
+      format: 'cjs',
+      file: './dist/index.cjs',
+      sourcemap: true,
+    },
+  ],
+  external: ['@lezer/lr'],
+  plugins: [lezer(), typescript({ tsconfig: './tsconfig.json' })],
+};
