+++ src/languages/satf/satf-sasf-utils.test.ts
@@ -1,4 +1,3 @@
-import { describe, expect, it } from 'vitest';
 import { satfToSequence } from './satf-sasf-utils';
 
 describe('satfToSequence', () => {
