+++ src/languages/seq-n/from-seq-json.ts
@@ -21,7 +21,7 @@ import type {
   VariableDeclaration,
 } from '@nasa-jpl/seq-json-schema/types';
 import { logError } from '../../logger';
-import { quoteEscape } from '../../sequence-utils';
+import { quoteEscape } from '../../utils/string';
 
 /**
  * Transform a sequence JSON time to it's sequence string form.
