+++ src/languages/seq-n/grammar.test.ts
@@ -1,6 +1,5 @@
 import { testTree } from '@lezer/generator/dist/test';
-import { describe, expect, test } from 'vitest';
-import { SeqLanguage } from './languages/seq-n/seq-n';
+import { SeqLanguage } from './seq-n';
 
 /*
  * Test cases here are canaries against unintended changes to the grammar. Certain grammar changes will alter the parse tree which
