--- src/languages/vml/vml.grammar.d.ts
@@ -1,2 +1,0 @@
-import { LRParser } from '@lezer/lr';
-export declare const parser: LRParser;
