+++ src/utils/highlight.ts
@@ -1,0 +1,3 @@
+import { Decoration } from '@codemirror/view';
+
+export const blockMark = Decoration.mark({ class: 'cm-block-match' });
