+++ src/utils/string.ts
@@ -1,0 +1,24 @@
+export function isQuoted(s: string): boolean {
+  return s.startsWith('"') && s.endsWith('"');
+}
+
+export function unquoteUnescape(s: string): string {
+  if (isQuoted(s) && s.length > 1) {
+    return s.slice(1, -1).replaceAll('\\"', '"');
+  }
+  return s;
+}
+
+export function quoteEscape(s: string): string {
+  return `"${s.replaceAll('"', '\\"')}"`;
+}
+
+export function removeEscapedQuotes(text: string): string;
+export function removeEscapedQuotes(text: number): number;
+export function removeEscapedQuotes(text: boolean): boolean;
+export function removeEscapedQuotes(text: string | number | boolean): string | number | boolean {
+  if (typeof text === 'string') {
+    return text.replace(/\\"|"(?!\\")/g, '"').trim();
+  }
+  return text;
+}
