+++ tsconfig.json
@@ -1,15 +1,15 @@
 {
   "compilerOptions": {
-    "module": "ESNext",
+    "module": "ES2022",
     "moduleResolution": "bundler",
     "target": "ESNext",
     "strict": true,
-    "verbatimModuleSyntax": true,
+    "verbatimModuleSyntax": false,
     "declaration": true,
     "sourceMap": true,
     "declarationMap": true,
     "esModuleInterop": true,
-    "rootDir": "src",
-    "outDir": "build"
+    // "rootDir": "src",
+    "outDir": "dist"
   }
 }
